+++ CHANGELOG.md
@@ -18,18 +18,67 @@
   </summary>
 
 ### :bug: Bug Fix
-  - `giphy`
-    - [#1840](https://github.com/wix-incubator/rich-content/pull/1840) prevent giphy crash on bad content
-</details>
+- `button`
+  - [#1860](https://github.com/wix-incubator/rich-content/pull/1860) fix `open in a new tab` radio button behaviour
 
+</details>
 <hr/>
+
+## 8.8.0 (Dec 13, 2020)
+### :rocket: New Feature
+- `theming`
+  - [#1862](https://github.com/wix-incubator/rich-content/pull/1862) quote border color
+
+### :bug: Bug Fix
+- `editor`
+  - [#1859](https://github.com/wix-incubator/rich-content/pull/1859) fix footerToolbarButtons height change after font properties changes (size, lineHeight, etc)
+- `general`
+  - [#1861](https://github.com/wix-incubator/rich-content/pull/1861) improve theming non-overrides to ensure non-breaking cssOverrides
+- `giphy`
+  - [#1840](https://github.com/wix-incubator/rich-content/pull/1840) prevent giphy crash on bad content
+- `image`
+  - [#1851](https://github.com/wix-incubator/rich-content/pull/1851) added rtl-ignore to fix the expand icon location
+
+### :book: Documentation
+- [#1856](https://github.com/wix-incubator/rich-content/pull/1856) README + documentation improved
+
+## 8.7.6 (Dec 11, 2020)
+### :bug: Bug Fix
+- `vertical embed`
+  - [#1855](https://github.com/wix-incubator/rich-content/pull/1855) Theme adaption
+
+## 8.7.5 (Dec 10, 2020)
+### :bug: Bug Fix
+- `gallery`
+  - [#1853](https://github.com/wix-incubator/rich-content/pull/1853) removed image titles from mobile galleries
+
+## 8.7.4 (Dec 10, 2020)
+### :bug: Bug Fix
+- `ricos-editor`
+  - [#1850](https://github.com/wix-incubator/rich-content/pull/1850) settings modal overlay appears above wix ad
+  
+### :house: Internal
+- `test-env`
+  - [#1848](https://github.com/wix-incubator/rich-content/pull/1848) theme `customStyles` coverage
+
+## 8.7.3 (Dec 10, 2020)
+### :rocket: New Feature
+- `editor`
+  - [#1830](https://github.com/wix-incubator/rich-content/pull/1830) conditional buttons
+
+## 8.7.2 (Dec 9, 2020)
+### :bug: Bug Fix
+- `giphy`
+  - [#1840](https://github.com/wix-incubator/rich-content/pull/1840) prevent giphy crash on bad content
+- `headers-markdown`
+  - [#1841](https://github.com/wix-incubator/rich-content/pull/1841) don't pull `wix-rich-content-plugin-commons` into viewer
 
 ## 8.7.1 (Dec 8, 2020)
 ### :house: Internal
-  - `gallery`
-    - [#1838](https://github.com/wix-incubator/rich-content/pull/1838) bump pro-gallery to 2.3.0
-  - `general`
-    - [#1836](https://github.com/wix-incubator/rich-content/pull/1836) moved to midgard-yarn
+- `gallery`
+  - [#1838](https://github.com/wix-incubator/rich-content/pull/1838) bump pro-gallery to 2.3.0
+- `general`
+  - [#1836](https://github.com/wix-incubator/rich-content/pull/1836) moved to midgard-yarn
 
 ## 8.7.0 (Dec 7, 2020)
 ### :rocket: New Feature